Added Player.class + style
In addition to adding a "Player" class I am also working on a new design style. Using the "this" when refering to variables in the class. I am evening doing this when it's not exactly needed, for example, when I am setting the value of the "score" variable to 0. Also, I am making my getters and setters one line. I used to not like this, but once I started working on larger projects (for school) I noticed that the 3 lines they took up made a big difference when you had more than 2 variables in a Class. This makes the code neater and easier to read/understand.